% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{df-mtx-convert}
\alias{df-mtx-convert}
\alias{mtx_df2mtx}
\alias{mtx_df2sparse_mtx}
\alias{mtx_mtx2df}
\alias{mtx_sparse_mtx2df}
\title{DataFrame and Matrix Conversion}
\usage{
mtx_df2mtx(df, row, col, value, na.values = 0)

mtx_df2sparse_mtx(df, row, col, value, na.values = 0)

mtx_mtx2df(mtx, row, col, value)

mtx_sparse_mtx2df(mtx, row, col, value, full = FALSE)
}
\arguments{
\item{df}{A long-format dataframe. It should contain at least 3 columns
specified by params \code{row}, \code{col}, and \code{value}.}

\item{row}{A string. The name of a column in \code{df} that contains the row
names of the matrix.}

\item{col}{A string. The name of a column in \code{df} that contains the column
names of the matrix.}

\item{value}{A string. The name of a column in \code{df} that contains the
values of the matrix.}

\item{na.values}{The value to replace NAs introduced during conversion.
Setting to \code{NA} or \code{NULL} to skip the value replacement.}

\item{mtx}{A matrix. It should be regular matrix for \code{mtx_mtx2df()} and
sparse matrix of \code{dgCMatrix} class for \code{mtx_sparse_mtx2df()}.}

\item{full}{A bool. Whether to return all combinations of rows and columns.}
}
\value{
\itemize{
\item The \code{mtx_df2mtx()} returns a matrix.
\item The \code{mtx_df2sparse_mtx()} returns a sparse matrix of \code{dgCMatrix} class.
\item The \code{mtx_mtx2df()} returns a long-format dataframe.
\item The \code{mtx_sparse_mtx2df()} returns a long-format dataframe.
}
}
\description{
The conversion between long-format dataframe and (sparse) matrix.
}
\section{Notes}{

The param \code{na.values} could affect the \code{NA}s that already exist in the
dataframe or matrix before conversion, which could lead to an unexpected
result.
}

\examples{
df <- data.frame(
  x = c("x1", "x1", "x2"),
  y = c("y1", "y2", "y2"),
  v = 1:3,
  stringsAsFactors = FALSE
)
mtx1 <- mtx_df2mtx(df, row = "x", col = "y", value = "v")
mtx2 <- mtx_df2sparse_mtx(df, row = "x", col = "y", value = "v")
df1 <- mtx_mtx2df(mtx1, row = "x", col = "y", value = "v")
df2 <- mtx_sparse_mtx2df(mtx2, row = "x", col = "y", value = "v")

}
