% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_ffprint}
\alias{io_ffprint}
\title{Force Flush Print}
\usage{
io_ffprint(s)
}
\arguments{
\item{s}{An object.}
}
\value{
Void.
}
\description{
Force Flush Print
}
\examples{
io_ffprint("hello world!")
}
\seealso{
\code{\link[=print]{print()}}
}
