% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io_write_tsv}
\alias{io_write_tsv}
\title{Write TSV}
\usage{
io_write_tsv(x, file)
}
\arguments{
\item{x}{A dataframe.}

\item{file}{A string. Path to the output TSV file.}
}
\value{
Void.
}
\description{
This function is aimed to write a dataframe into a TSV file in a
commonly used format.
Specifically, it calls \code{[utils]{write.table}} internally with
parameters \code{sep = "\\t"}, \code{quote = FALSE}, \code{row.names = FALSE}, and
\code{col.names = TRUE}.
}
\examples{
\dontrun{df <- data.frame(x = 1:5, y = 2:6)
io_write_tsv(df, "~/test_rrbio/io_write_tsv.tsv")}
}
