% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{mtx_mtx2dgCMatrix}
\alias{mtx_mtx2dgCMatrix}
\title{Convert a Matrix to Sparse Matrix of dgCMatrix Class
Notes: it seems sparse matrix of many classes can be successfully converted
to \code{dgTMatrix} while fail to be converted to \code{dgCMatrix} directly.
Therefore, x -> dgTMatrix -> dgCMatrix could be safer for format
conversion.}
\usage{
mtx_mtx2dgCMatrix(x)
}
\description{
Convert a Matrix to Sparse Matrix of dgCMatrix Class
Notes: it seems sparse matrix of many classes can be successfully converted
to \code{dgTMatrix} while fail to be converted to \code{dgCMatrix} directly.
Therefore, x -> dgTMatrix -> dgCMatrix could be safer for format
conversion.
}
\keyword{internal}
