% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os_assert_e}
\alias{os_assert_e}
\title{Assert Existencce of File or Directory}
\usage{
os_assert_e(path, type = "file")
}
\arguments{
\item{path}{A string. Path to a file or directory.}

\item{type}{A string. The path type, "\code{file}" or "\code{dir}".}
}
\value{
Void.
}
\description{
This function is aimed to assert the existence of specified file or
directory.
}
\section{Details}{

It will call \code{\link[base:stop]{stop()}} and raise an error if the path
does not exist.
}

\examples{
\dontrun{
os_assert_e("~/test")
os_assert_e("~/debug/test.R")}
}
