% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os_join_path}
\alias{os_join_path}
\title{Join Two Path}
\usage{
os_join_path(path1, path2)
}
\arguments{
\item{path1}{A string. The first path.}

\item{path2}{A string. The second path.}
}
\value{
A string. Concatenated path.
}
\description{
This function is aimed to concatenate two path. It should be commonly
used to concatenate a directory and a file path to generate the absolute
path of the file.
}
\examples{
dir_path <- "~/debug"
file_path <- "test.R"
full_path <- os_join_path(dir_path, file_path)  # "~/debug/test.R"
}
