% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os_safe_mkdir}
\alias{os_safe_mkdir}
\title{Create Directory}
\usage{
os_safe_mkdir(dir, recursive = TRUE)
}
\arguments{
\item{dir}{A string. Path to the directory.}

\item{recursive}{A bool. Whether to create the sub-directory recursively.}
}
\value{
Void.
}
\description{
This function is aimed to create a directory if it does not exist.
}
\examples{
\dontrun{os_safe_mkdir("~/debug/test1/test2")}
}
