% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pkg_install}
\alias{pkg_install}
\title{Package Installation}
\usage{
pkg_install(
  pkgs,
  ...,
  lib = NULL,
  update = FALSE,
  ask = FALSE,
  verbose = TRUE,
  max_try = 3
)
}
\arguments{
\item{pkgs}{\code{character()} vector of package names to install or
update.  A missing value updates installed packages according
to \verb{update =} and \verb{ask =}. Package names containing a '/' are
treated as GitHub repositories and installed using
\code{remotes::install_github()}.}

\item{...}{Additional arguments used by \code{BiocManager::install()} and
\code{utils::install.packages()}.}

\item{lib}{
    character vector giving the library directories where to
    install the packages.  Recycled as needed.  If missing, defaults to
    the first element of \code{\link{.libPaths}()}.
  }

\item{update}{\code{logical(1)}. When \code{FALSE}, \code{BiocManager::install()}
does not attempt to update old packages. When \code{TRUE}, update
old packages according to \code{ask}.}

\item{ask}{\code{logical(1)} indicating whether to prompt user before
installed packages are updated.  If TRUE, user can choose
whether to update all outdated packages without further
prompting, to pick packages to update, or to cancel updating
(in a non-interactive session, no packages will be updated
unless \code{ask = FALSE}).}

\item{verbose}{
    a logical indicating if some \dQuote{progress report} should be given.
  }

\item{max_try}{An integer. The maximum rounds of installation.}
}
\value{
A vector of names of packages whose installation failed.
}
\description{
This function is a wrapper of BiocManager::install(), allowing several
rounds of installation in case of errors during installation. Note that
BiocManager::install() can also install CRAN and GitHub packages, in
addition to Bioconductor packages.
}
\section{Notes}{

This wrapper function changes the default behavior of several parameters:
\itemize{
\item The \code{lib}: default is \code{NULL}, which will use the first element of
\code{\link[=.libPaths]{.libPaths()}}.
\item The \code{update}: default is \code{FALSE}, which will not update the old packages.
\item The \code{ask}: default is \code{FALSE}, which will not prompt user before
installed packages are updated.
}
}

\examples{
\dontrun{
# install CRAN and Bioconductor packages
pkgs <- c("dplyr", "GenomicRanges")
pkg_install(pkgs)
}
}
