% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrbio-package.R
\docType{package}
\name{rrbio-package}
\alias{rrbio}
\alias{rrbio-package}
\title{rrbio: Handy R Routine Functions for Bioinformatics Analysis}
\description{
Handy R Routine Functions for Bioinformatics Analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hxj5/rrbio}
  \item Report bugs at \url{https://github.com/hxj5/rrbio/issues}
}

}
\author{
\strong{Maintainer}: Xianjie Huang \email{xianjie5@connect.hku.hk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
