% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_count.R
\name{spcnt-mtx-io}
\alias{spcnt-mtx-io}
\alias{spcnt_load_matrices}
\alias{spcnt_save_matrices}
\title{Spliced and Unspliced Count Matrix Input/Output}
\usage{
spcnt_load_matrices(
  in_dir,
  spliced_fn,
  unspliced_fn,
  ambiguous_fn = NULL,
  gene_fn = NULL,
  cell_fn = NULL,
  gene_header = FALSE,
  cell_header = FALSE,
  gene_index = 1,
  cell_index = 1
)

spcnt_save_matrices(
  out_dir,
  spliced_mtx,
  spliced_fn,
  unspliced_mtx,
  unspliced_fn,
  ambiguous_mtx = NULL,
  ambiguous_fn = NULL,
  gene_fn = NULL,
  cell_fn = NULL
)
}
\arguments{
\item{in_dir}{Path to the input directory.}

\item{spliced_fn}{A string. Path to the spliced matrix file.}

\item{unspliced_fn}{A string. Path to the unspliced matrix file.}

\item{ambiguous_fn}{A string. Path to the ambiguous matrix file.}

\item{gene_fn}{A string. Path to the gene annotation file.}

\item{cell_fn}{A string. Path to the cell annotation file.}

\item{gene_header}{A bool. Whether \code{gene_fn} has header.}

\item{cell_header}{A bool. Whether \code{cell_fn} has header.}

\item{gene_index}{An integer or a string. The index (integer) or name
(string) of the column in \code{gene_fn} that stores the annotation.}

\item{cell_index}{An integer or a string. The index (integer) or name
(string) of the column in \code{cell_fn} that stores the annotation.}

\item{out_dir}{A string. Path to the output directory.}

\item{spliced_mtx}{The \verb{gene x cell} spliced count matrix.}

\item{unspliced_mtx}{The \verb{gene x cell} unspliced count matrix.}

\item{ambiguous_mtx}{The \verb{gene x cell} ambiguous count matrix.}
}
\value{
\itemize{
\item The \code{spcnt_load_matrices()} returns a list of five elements: (1) \code{spliced}:
the spliced count matrix; (2) \code{unspliced}: the unspliced count matrix; (3)
\code{ambiguous}: the ambiguous count matrix; (4) \code{gene_anno}: a dataframe, the
annotation of genes; (5) \code{cell_anno}: a dataframe, the annotations of
cells. Note that \code{spliced}, \code{unspliced}, and \code{ambiguous} are all of
\code{dgCMatrix} class.
\item The \code{spcnt_save_matrices()} returns Void.
}
}
\description{
These functions are aimed to read and write spliced and unspliced UMI/read
count matrices generated from scRNA-seq data.
}
\examples{
\dontrun{
genes <- paste0("gene", 1:5)
cells <- paste0("cell", 1:4)

spliced_mtx <- matrix(rpois(20, 1), nrow = 5)
rownames(spliced_mtx) <- genes
colnames(spliced_mtx) <- cells

unspliced_mtx <- matrix(rpois(20, 1), nrow = 5)
rownames(unspliced_mtx) <- genes
colnames(unspliced_mtx) <- cells

ambiguous_mtx <- matrix(0, nrow = 5, ncol = 4)
rownames(ambiguous_mtx) <- genes
colnames(ambiguous_mtx) <- cells

spcnt_save_matrices(
  "~/test_rrbio", 
  spliced_mtx, "spliced.mtx",
  unspliced_mtx, "unspliced.mtx",
  ambiguous_mtx, "ambiguous.mtx",
  "genes.tsv", "barcodes.tsv"
)

res <- spcnt_load_matrices(
  "~/test_rrbio", "spliced.mtx", "unspliced.mtx", "ambiguous.mtx",
  "genes.tsv", "barcodes.tsv"
)
}

}
\seealso{
\code{\link[=mtx_load_sparse_mtx_n]{mtx_load_sparse_mtx_n()}} and \code{\link[=mtx_save_sparse_mtx_n]{mtx_save_sparse_mtx_n()}}
}
