% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time-str}
\alias{time-str}
\alias{time_str_date}
\alias{time_str_time}
\alias{time_str_datetime}
\alias{time_str_datetime0}
\alias{time_strftime}
\title{Format Time String}
\usage{
time_str_date(time = NULL, sep = "")

time_str_time(time = NULL, sep = ":")

time_str_datetime(time = NULL)

time_str_datetime0(time = NULL)

time_strftime(time = NULL, format = NULL)
}
\arguments{
\item{time}{A \code{POSIXct} object. Setting to \code{NULL} to use current time.}

\item{sep}{A string. Delimiter to join the sub-string in \code{format}.}

\item{format}{A string. A \code{format} string, e.g., "\verb{\%Y-\%m-\%d \%H:\%M:\%S}".}
}
\value{
Every function returns a string.
}
\description{
These functions are aimed to format time string.
}
\section{Notes}{

The \code{format} string for different functions:
\itemize{
\item for \code{time_str_date()}: "\verb{\%Y}", "\verb{\%m}", "\verb{\%d}" joined by \code{sep}.
\item for \code{time_str_time()}: "\verb{\%H}", "\verb{\%M}", "\verb{\%S}" joined by \code{sep}.
\item for \code{time_str_datetime()}: "\verb{\%Y-\%m-\%d \%H:\%M:\%S}".
\item for \code{time_str_datetime0()}: "\verb{\%Y\%m\%d \%H:\%M:\%S}".
\item for \code{time_strftime()}: specified by user. Default is "\verb{\%Y-\%m-\%d \%H:\%M:\%S}"
when \code{format} is set to \code{NULL}.
}
}

\examples{
p <- as.POSIXct("2023-09-19 10:00:00")
time_str_date(p)       # "20230919"
time_str_date(p, sep = "-")  # "2023-09-19"
time_str_time(p)       # "10:00:00"

# use current time
time_str_date()  
time_str_time()  
time_str_datetime()  
time_str_datetime0()

}
\seealso{
\link{POSIXct} and \code{\link[base:format]{base::format()}}
}
