% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{spmtx-io}
\alias{spmtx-io}
\alias{mtx_load_sparse_mtx}
\alias{mtx_save_sparse_mtx}
\title{Sparse Matrix Input/Output}
\usage{
mtx_load_sparse_mtx(
  in_dir,
  mtx_fn,
  row_fn = NULL,
  col_fn = NULL,
  row_header = FALSE,
  col_header = FALSE
)

mtx_save_sparse_mtx(mtx, out_dir, mtx_fn, row_fn = NULL, col_fn = NULL)
}
\arguments{
\item{in_dir}{A string. Path to the input directory.}

\item{mtx_fn}{A string. Path to the file storing the sparse matrix, in
Matrix Market Format.}

\item{row_fn}{A string. Path to the file storing the annotation of matrix
rows. Setting to NULL if there is no such annotation.}

\item{col_fn}{A string. Path to the file storing the annotation of matrix
columns. Setting to NULL if there is no such annotation.}

\item{row_header}{A bool. Whether the row annotation file has header.}

\item{col_header}{A bool. Whether the column annotation file has header.}

\item{mtx}{A matrix. It could be either sparse matrix or regular matrix.}

\item{out_dir}{A string. Path to the output directory.}
}
\value{
\itemize{
\item The \code{mtx_load_sparse_mtx()} returns a sparse matrix of \code{dgCMatrix} class.
\item The \code{mtx_save_sparse_mtx()} returns Void.
}
}
\description{
These functions are aimed to load and save the sparse matrices in
Matrix Market Format, together with its annotation files.
}
\section{Notes}{

The matrix file, row annotation file (if available) and column annotation
file (if available) should be in the same directory.

The row and column names of the matrix should be the first column of the
corresponding annotation file.
}

\examples{
m <- matrix(rpois(20, 1), nrow = 5)
rownames(m) <- paste0("x", 1:5)
colnames(m) <- paste0("y", 1:4)
mtx_save_sparse_mtx(m, "~/tmp_rrbio", "matrix.mtx", "rows.tsv", "cols.tsv")
m2 <- mtx_load_sparse_mtx("~/tmp_rrbio", "matrix.mtx", "rows.tsv", "cols.tsv")

}
